Use fp32 atomicAdd in one mode kernel.
It's faster than fp16 atomicAdd, don't know exactly why.
Since it has both better performance and accuracy, I'm using it for all
arch.

+++ rwkv_pip_package/src/rwkv/cuda/wrapper.cpp
@@ -15,7 +15,7 @@ void cuda_mm8_one(int N, int M,
                   uint8_t *w, int w_stride,
                   fp16 *mx, fp16 *rx,
                   fp16 *my, fp16 *ry,
-                  fp16 *y);
+                  float *y);
 
 void wkv_forward(int64_t B, int64_t T, int64_t C, torch::Tensor &w, torch::Tensor &u, torch::Tensor &k, torch::Tensor &v, torch::Tensor &y, torch::Tensor &aa, torch::Tensor &bb, torch::Tensor &pp) {
     cuda_wkv_forward(B, T, C, w.data_ptr<float>(), u.data_ptr<float>(), k.data_ptr<fp16>(), v.data_ptr<fp16>(), y.data_ptr<fp16>(), aa.data_ptr<float>(), bb.data_ptr<float>(), pp.data_ptr<float>());
@@ -54,7 +54,7 @@ void mm8_one(int64_t N, int64_t M,
         w.data_ptr<uint8_t>(), w.stride(0),
         mx.data_ptr<fp16>(), rx.data_ptr<fp16>(),
         my.data_ptr<fp16>(), ry.data_ptr<fp16>(),
-        y.data_ptr<fp16>());
+        y.data_ptr<float>());
 }
 
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
